Simplify free suffix match support
Do not use the free_suffix variable directly in the mask generation code,
instead set the is_char field to None when the domain part length needs
to be masked.

This fixes a bug that caused different BPF code to be generated for the
www** rule if another non-free suffix rule was also supplied.

The following is correct:

    $ ./bpfgen -sn dns -- www**
    ...
    lb_0:
        ; ldx M[0]
        ; Match: 03777777 '\x03www'  mask=ff000000
        ld [x + 0]
        or #0xff000000
        jneq #0xff777777, lb_1
        ret #0
    ...

While the following is not:

    $ ./bpfgen -sn dns -- www** www.example.com
    ...
    lb_0:
        ; ldx M[0]
        ; Match: 03777777 '\x03www'  mask=00000000
        ld [x + 0]
        jneq #0x03777777, lb_1
        ret #0
    ...

(note in particular the different masks generated).