Fix free suffix match with subdomain
The www.** and www** rules were equivalent, meaning that www.** matched
both www.example.com and wwwxxx.example.com, which is wrong. To fix this
we avoid masking the last domain part length to make sure the provided
domain part is matched entirely.

This changes the output from the following (wrong):

    $ ./bpfgen -sn dns -- www** www.**
    ...
    lb_0:
        ; ldx M[0]
        ; Match: 03777777 '\x03www'  mask=ff000000
        ld [x + 0]
        or #0xff000000
        jneq #0xff777777, lb_1
        ret #0

    lb_1:
        ldx M[0]
        ; Match: 03777777 '\x03www'  mask=ff000000
        ld [x + 0]
        or #0xff000000
        jneq #0xff777777, lb_2
        ret #0
    ...

To the following (correct):

    $ ./bpfgen -sn dns -- www** www.**
    ...
    lb_0:
        ; ldx M[0]
        ; Match: 03777777 '\x03www'  mask=ff000000
        ld [x + 0]
        or #0xff000000
        jneq #0xff777777, lb_1
        ret #0

    lb_1:
        ldx M[0]
        ; Match: 03777777 '\x03www'  mask=00000000
        ld [x + 0]
        jneq #0x03777777, lb_2
        ret #0
    ...

(note in particular the different masks generated for lb_1).